chore: rename icon to old icon
