+++ core/blockly.ts
@@ -69,7 +69,7 @@ import {VerticalFlyout} from './flyout_vertical.js';
 import {CodeGenerator} from './generator.js';
 import {Gesture} from './gesture.js';
 import {Grid} from './grid.js';
-import {Icon} from './icon.js';
+import {Icon} from './icon_old.js';
 import {inject} from './inject.js';
 import {Align, Input} from './inputs/input.js';
 import {inputTypes} from './inputs/input_types.js';
