+++ core/renderers/measurables/icon.ts
@@ -8,7 +8,7 @@ import * as goog from '../../../closure/goog/goog.js';
 goog.declareModuleId('Blockly.blockRendering.Icon');
 
 /* eslint-disable-next-line no-unused-vars */
-import type {Icon as BlocklyIcon} from '../../icon.js';
+import type {Icon as BlocklyIcon} from '../../icon_old.js';
 import type {ConstantProvider} from '../common/constants.js';
 
 import {Measurable} from './base.js';
